add edit functionallity

+++ controllers/playController.js
@@ -56,8 +56,56 @@ router.get('/details/:id', async (req, res) => {
     }
 });
 
-router.get('/delete/:id', async (req, res) => {
+router.get('/edit/:id', isUser(), async (req, res) => {
     try {
+        const play = await req.storage.getPlayById(req.params.id);
+
+        if (play.author != req.user._id) {
+            throw new Error('Cannot edit play you haven\'t created!');
+        }
+
+        res.render('play/edit', { play });
+    } catch (err) {
+        console.log(err.message);
+        res.redirect('/play/details' + req.params.id);
+    }
+});
+
+router.post('/edit/:id', isUser(), async (req, res) => {
+    try {
+        const play = await req.storage.getPlayById(req.params.id);
+
+        if (play.author != req.user._id) {
+            throw new Error('Cannot edit play you haven\'t created!');
+        }
+
+        await req.storage.editPlay(req.params.id, req.body);
+
+        res.redirect('/');
+    } catch (err) {
+        const ctx = {
+            errors: parseError(err),
+            play: {
+                _id: req.params.id,
+                title: req.body.title,
+                description: req.body.description,
+                imageUrl: req.body.imageUrl,
+                public: Boolean(req.body.public)
+            }
+        }
+        res.render('play/edit', ctx);
+    }
+    
+});
+
+router.get('/delete/:id', isUser(), async (req, res) => {
+    try {
+        const play = await req.storage.getPlayById(req.params.id);
+
+        if (play.author != req.user._id) {
+            throw new Error('Cannot delete play you haven\'t created!');
+        }
+
         await req.storage.deletePlay(req.params.id)
         res.redirect('/');
     } catch (err) {
